Delete warehouse address and bank details popup Implemented

See merge request commerce/frontend/seller!37

+++ src/app/views/user/profile/address-profile/address-profile.component.ts
@@ -3,6 +3,8 @@ import { Address } from 'src/app/shared/models/address';
 import { ActivatedRoute, ActivatedRouteSnapshot, Router } from '@angular/router';
 import { UserService } from 'src/app/shared/services/user.service';
 import { Subject, Subscription } from 'rxjs';
+import { MatDialog } from '@angular/material';
+import { DeleteWarehouseAddressComponent } from 'src/app/shared/dialogs/delete-warehouse-address/delete-warehouse-address.component';
 
 @Component({
     selector: 'app-address-profile',
@@ -12,7 +14,8 @@ export class AddressProfileComponent implements OnInit, OnDestroy {
     selectedProfile: any;
     constructor(private activatedRoute: ActivatedRoute,
         private userService: UserService,
-        private router: Router) { }
+        private router: Router,
+        private _dialog: MatDialog) { }
 
     addresses: Address[];
     addressType: string;
@@ -29,7 +32,6 @@ export class AddressProfileComponent implements OnInit, OnDestroy {
                 this.userService.enableProfile$.next(true);
             }
         });
-        //console.log(this.selectedProfile);
         this.startSubscriptions();
     }
 
@@ -69,9 +71,33 @@ export class AddressProfileComponent implements OnInit, OnDestroy {
     }
 
     deleteAddrs(addressId: number): void {
-        this.userService.deleteAddress(addressId).then(res => {
-            const deletedIndex = this.addresses.findIndex(addrs => addrs.addressId == addressId);
-            this.addresses.splice(deletedIndex, 1);
-        });
+
+        if(this.addressType === 'warehouse'){
+            /**
+             * Open popup if deleted address is warehouse address
+             */
+            const d = this._dialog.open(DeleteWarehouseAddressComponent, {
+                data: { addressId: addressId, addresses: this.addresses },
+                disableClose: true,
+                panelClass: 'profile-verification-popup',
+                width: '25%'
+            });
+            d.afterClosed().toPromise().then((data: any) => {
+                if (data) {
+                    this.addresses = data;
+                    this.userService.getUserPercentage().then(res => this.userService.updatePercentage$.next(res));
+                }
+            });
+        }else{
+
+            /**
+             * No popup when deleted address is billing address
+             */
+            this.userService.deleteAddress(addressId).then(res => {
+                const deletedIndex = this.addresses.findIndex(addrs => addrs.addressId == addressId);
+                this.addresses.splice(deletedIndex, 1);
+                this.userService.getUserPercentage().then(res => this.userService.updatePercentage$.next(res));
+            });
+        }   
     }
 }
